requirements.txt
finnhub_python==2.4.19
Flask==3.0.1
Flask_SocketIO==5.3.6
keras==2.13.1
matplotlib==3.7.4
nltk==3.8.1
numpy==1.23.5
pandas==2.0.3
pika==1.3.2
tensorflow==2.13.0
tensorflow_intel==2.13.0
transformers==4.37.1
yfinance==0.2.36
torch==2.1.1
